+++ jacquard/variant_caller_transforms/strelka.py
@@ -23,7 +23,7 @@ import jacquard.utils.vcf as vcf
 
 JQ_STRELKA_TAG = "JQ_SK_"
 STRELKA_ABBREVIATION = "SK"
-VERSION = "v1.0.14"
+VERSION = "v1.0-2.9"
 
 class _GenotypeTag(common_tags.AbstractJacquardTag):
     #pylint: disable=too-few-public-methods
@@ -379,4 +379,3 @@ class _StrelkaVcfReader(object):
         for tag in self.tags:
             tag.add_tag_values(vcf_record)
         return vcf_record
-